+++ source/ast/types/AllTypes.cpp
@@ -693,7 +693,7 @@ const Type& FixedSizeUnpackedArrayType::fromDim(const Scope& scope, const Type& 
 
     auto& comp = scope.getCompilation();
     auto selectableWidth = checkedMulU32(elementType.getSelectableWidth(), dim.width());
-    auto bitstreamWidth = checkedMulU32((uint32_t)elementType.getBitstreamWidth(), dim.width());
+    auto bitstreamWidth = checkedMulU32(elementType.getBitstreamWidth(), dim.width());
 
     if (!selectableWidth || selectableWidth > MaxBitWidth || !bitstreamWidth ||
         bitstreamWidth > MaxBitWidth) {
@@ -1053,7 +1053,7 @@ const Type& UnpackedUnionType::fromSyntax(const ASTContext& context,
             result->selectableWidth = std::max(result->selectableWidth,
                                                field->getType().getSelectableWidth());
             result->bitstreamWidth = std::max(result->bitstreamWidth,
-                                              (uint32_t)field->getType().getBitstreamWidth());
+                                              field->getType().getBitstreamWidth());
         }
     }
